+++ src/types/actor.d.ts
@@ -190,3 +190,10 @@ export interface DivineData {
 
 
 
+
+
+
+
+
+
+