+++ src/types/item.d.ts
@@ -140,3 +140,10 @@ export interface ArmorData {
 
 
 
+
+
+
+
+
+
+