+++ src/utils/calculations.ts
@@ -175,3 +175,10 @@ export function validateSkillValue(skillValue: number, masteryRank: number): num
 
 
 
+
+
+
+
+
+
+